# rumicoach-backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: rumi.coach API
  version: 1.0.0
servers:
  - url: /v1
paths:
...
  /me/profile:
    get:
      summary: Get user profile and progress statistics
      tags:
        - Users
      operationId: GetMeProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Unauthorized
...
  /memories:
    get:
      summary: Get user memories
      tags:
        - Users
      operationId: GetMemories
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/MemoryCategory'
        - name: text
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of user memories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryPaginatedResponse'
        '401':
          description: Unauthorized
  /memories/{memoryId}:
    delete:
      summary: Delete a specific memory
      tags:
        - Users
      operationId: DeleteMemory
      security:
        - BearerAuth: []
      parameters:
        - name: memoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Memory not found
...
  /journey:
    get:
      summary: Get today's proposed growth session
      tags:
        - DailyGrowth
      operationId: GetDailyJourney
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Session proposal for the day (can be null if none)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneySessionResponse'
        '401':
          description: Unauthorized
  /quotes/random:
    get:
      summary: Get a random personal development quote
      tags:
        - Quotes
      operationId: GetRandomQuote
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/QuoteCategory'
      responses:
        '200':
          description: A random quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
...
  /recommendations:
    get:
      summary: Get user recommendations
      tags:
        - Users
      operationId: GetRecommendations
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [book, article, video, podcast, other]
      responses:
        '200':
          description: Paginated list of user recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationPaginatedResponse'
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Memory:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        category:
          $ref: '#/components/schemas/MemoryCategory'
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    MemoryPaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Memory'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    MemoryCategory:
      type: string
      enum:
        - identity
        - values
        - needs
        - context
        - obstacles
        - insight
...
    PaginationInfo:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
...
    JourneySessionResponse:
      type: object
      properties:
        session:
          $ref: '#/components/schemas/SessionType'
          nullable: true
        mode:
          type: string
          enum:
            - start
            - resume
          nullable: true
        quote:
          $ref: '#/components/schemas/Quote'
          nullable: true
        theme:
          type: string
          nullable: true
          description: "Visual theme override picked by the AI at the end of the last session. Only present while that override is active — absent when the user's own theme choice (on /me, PATCH /me clears the override) should stand. Clients must not change the theme when this is absent."
        nextSession:
          $ref: '#/components/schemas/SessionInfo'
          nullable: true
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionInfo'
          nullable: true
          description: "List of upcoming deep sessions on the user's growth journey and their estimated availability dates."
        focusArea:
          type: string
          nullable: true
          description: "The user's Current Focus — the Wheel of Life area being prioritized (chosen during onboarding), in the user's language."
    SessionInfo:
      type: object
      description: "The next deep session on the user's growth journey and when it becomes available, so the app can preview the upcoming chapter."
      properties:
        session:
          $ref: '#/components/schemas/SessionType'
        availableAt:
          type: string
          format: date-time
          description: "When the session unlocks. May be in the past, meaning it is available now."
      required:
        - session
        - availableAt
    SessionType:
      type: string
      enum:
        - onboarding
        - session_vision
        - session_movement
        - session_values
        - session_energy
        - session_decisions
        - session_beliefs
        - session_identity
        - session_acceptance
        - session_priorities
        - checkin
    Quote:
      type: object
      properties:
        id:
          type: string
        quote:
          type: string
        author:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/QuoteCategory'
    QuoteCategory:
      type: string
      enum:
        - growth
        - resilience
        - action
        - mindset
        - wisdom
        - purpose
...
    Recommendation:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        sessionId:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [book, article, video, podcast, other]
        author:
          type: string
        url:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    RecommendationPaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    ProfileVision:
      type: object
      properties:
        text:
          type: string
          nullable: true
        craftedAt:
          type: string
          format: date-time
          nullable: true
    ProfileBalance:
      type: object
      properties:
        completedAt:
          type: string
          format: date
          nullable: true
        data:
          type: string
          nullable: true
          description: "JSON array of WheelOfLifeItem"
    ProfileProgress:
      type: object
      properties:
        currentStreak:
          type: integer
        bestStreak:
          type: integer
        totalSessions:
          type: integer
        hoursWithRumi:
          type: number
          format: float
        insightsDiscovered:
          type: integer
        commitmentsKept:
          type: integer
    BadgeType:
      type: string
      description: >-
        Profile badges v2. The app renders these in this order as a 3x5 grid.
        Rows in user_badges may still carry the retired goalReached type (granted
        on behaviour-plan graduation) — kept for analytics; the app ignores types
        it does not recognize.
      enum:
        - firstSession
        - visionSet
        - firstCommitment
        - firstDeepSession
        - alwaysWithYou
        - threeDayStreak
        - tenInsights
        - allThemesExplored
        - sevenDayStreak
        - twentySessions
        - twentyFiveCommitments
        - thirtyDayStreak
        - wheelRemapped
        - areaImproved
        - hundredSessions
    ProfileBadge:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/BadgeType'
        earnedAt:
          type: string
          format: date-time
    UserProfileResponse:
      type: object
      properties:
        focusArea:
          type: string
          description: The user's chosen Focus Area for their coaching journey.
        vision:
          $ref: '#/components/schemas/ProfileVision'
        lifeBalance:
          $ref: '#/components/schemas/ProfileBalance'
        progress:
          $ref: '#/components/schemas/ProfileProgress'
        badges:
          type: array
          items:
            $ref: '#/components/schemas/ProfileBadge'
...